requirements.txt
streamlit
pdfplumber
python-docx
scikit-learn
spacy
pandas
